+++ jpsonic-main/src/main/java/com/tesshu/jpsonic/dao/InternetRadioDao.java
@@ -21,12 +21,13 @@
 
 package com.tesshu.jpsonic.dao;
 
-import static com.tesshu.jpsonic.dao.DaoUtils.nullableInstantOf;
+import static com.tesshu.jpsonic.dao.base.DaoUtils.nullableInstantOf;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.List;
 
+import com.tesshu.jpsonic.dao.base.TemplateWrapper;
 import com.tesshu.jpsonic.domain.InternetRadio;
 import org.checkerframework.checker.nullness.qual.Nullable;
 import org.slf4j.Logger;
