+++ jpsonic-main/src/main/java/com/tesshu/jpsonic/dao/TranscodingDao.java
@@ -21,12 +21,13 @@
 
 package com.tesshu.jpsonic.dao;
 
-import static com.tesshu.jpsonic.dao.DaoUtils.questionMarks;
+import static com.tesshu.jpsonic.dao.base.DaoUtils.questionMarks;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.List;
 
+import com.tesshu.jpsonic.dao.base.TemplateWrapper;
 import com.tesshu.jpsonic.domain.Transcoding;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.stereotype.Repository;
