+++ jpsonic-main/src/main/java/com/tesshu/jpsonic/dao/base/LegacyHsqlDaoHelper.java
@@ -19,7 +19,7 @@
  * (C) 2018 tesshucom
  */
 
-package com.tesshu.jpsonic.dao;
+package com.tesshu.jpsonic.dao.base;
 
 import java.sql.Connection;
 import java.sql.SQLException;
