+++ jpsonic-main/src/test/java/com/tesshu/jpsonic/AbstractNeedsScan.java
@@ -29,9 +29,9 @@ import java.util.function.Supplier;
 
 import javax.annotation.PostConstruct;
 
-import com.tesshu.jpsonic.dao.DaoHelper;
 import com.tesshu.jpsonic.dao.MusicFolderDao;
 import com.tesshu.jpsonic.dao.StaticsDao;
+import com.tesshu.jpsonic.dao.base.DaoHelper;
 import com.tesshu.jpsonic.service.MediaScannerService;
 import com.tesshu.jpsonic.service.SecurityService;
 import com.tesshu.jpsonic.service.ServiceMockUtils;
