+++ jpsonic-main/src/test/java/com/tesshu/jpsonic/dao/InternetRadioDaoTest.java
@@ -27,6 +27,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import java.time.temporal.ChronoUnit;
 
 import com.tesshu.jpsonic.NeedsHome;
+import com.tesshu.jpsonic.dao.base.GenericDaoHelper;
 import com.tesshu.jpsonic.domain.InternetRadio;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
