+++ jpsonic-main/src/test/java/com/tesshu/jpsonic/dao/PodcastDaoTest.java
@@ -31,6 +31,7 @@ import java.time.temporal.ChronoUnit;
 import java.util.List;
 
 import com.tesshu.jpsonic.NeedsHome;
+import com.tesshu.jpsonic.dao.base.GenericDaoHelper;
 import com.tesshu.jpsonic.domain.PodcastChannel;
 import com.tesshu.jpsonic.domain.PodcastEpisode;
 import com.tesshu.jpsonic.domain.PodcastStatus;
